% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listDatasets.R
\name{listDatasets}
\alias{listDatasets}
\title{List all datasets}
\usage{
listDatasets()
}
\value{
A \linkS4class{DataFrame}, containing the following fields
\itemize{
\item \code{Title}, short name of this data.
\item \code{Description}, description of the data.
\item \code{Omics}, omic layers profiled in the data.
\item \code{Species}, species of the data.
\item \code{RDataPath}, the corresponding rds files in this package.
}
}
\description{
This lists the summary information for all available datasets in the \pkg{TransOmicsData} package.
}
\details{
This package contains datasets spanning various biological contexts such as in vitro embryonic and tissue-specific development in mouse and human extracted from different sequencing technologies.
}
\examples{
listDatasets()

}
\author{
Carissa Chen
}
